# Conflicts:
#	net.jadoth.persistence.binary/src/net/jadoth/persistence/binary/internal/AbstractBinaryHandlerNative.java
#	net.jadoth.persistence.binary/src/net/jadoth/persistence/binary/internal/AbstractBinaryHandlerNativeCustom.java
#	net.jadoth.persistence.binary/src/net/jadoth/persistence/binary/internal/AbstractBinaryHandlerNativeCustomValue.java
#	net.jadoth.persistence.binary/src/net/jadoth/persistence/binary/internal/AbstractBinaryHandlerNativeCustomValueFixedLength.java

+++ net.jadoth.core/src/net/jadoth/collections/XUtilsArrayCollection.java
@@ -474,7 +474,7 @@ public final class XUtilsArrayCollection
 	}
 
 	public static final <E, A extends AbstractSimpleArrayCollection<E> & XSettingList<E>>
-	int rngReplace(
+	long rngReplace(
 		final A a,
 		final int offset,
 		final int length,
